refatorizar o codigo pra usar apontadores inves de ID
como o prof queixou-se que IDs pra ligar instancias estava mal segundo os conceitos de POO, meti apontadores nas classes modelo (ex: `Categoria* categoria` em vez de `int id_categoria` no Produto).
O repo tbm agr gere os apontadores com delete pa nao dar leak e tudo cria com new.
Aproveitei pra trocar nullptrs por NULL pa ficar mais parecido ao style do stor e arrumei o warning do size().

+++ Project/src/view/CatalogoView.cpp
@@ -31,7 +31,7 @@ void CatalogoView::printListaProdutos(const std::vector<ProdutoDTO>& lista) {
         return;
     }
     std::cout << "\n--- LISTA DE PRODUTOS ---" << std::endl;
-    for (int i = 0; i < (int)lista.size(); i++) {
+    for (size_t i = 0; i < lista.size(); i++) {
         std::cout << "ID: " << lista[i].id
                   << " | Nome: " << lista[i].nome
                   << " | Preco: " << lista[i].preco
